requirements.txt
fastapi==0.115.5
starlette==0.41.3
PyJWT==2.10.1
cryptography>=42.0.0
python-multipart>=0.0.18
sqlmodel==0.0.21
uvicorn>=0.30.5
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.2.1
pydantic>=2.8.2
email-validator>=2.0.0
